services/optimizer-agents/requirements-optimizer.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
structlog==23.2.0

# Kafka
confluent-kafka==2.3.0
avro-python3==1.10.2

# gRPC - versão consolidada
grpcio>=1.75.1
grpcio-tools>=1.75.1
grpcio-health-checking>=1.75.1
protobuf>=5.27.0

# Databases
pymongo==4.6.0
motor==3.3.2
redis[hiredis]>=5.0.1
neo4j==5.14.0

# ML/Optimization
dowhy==0.10.1
gymnasium==0.29.1
prophet==1.1.5
statsmodels==0.14.0
clickhouse-driver==0.2.6
pmdarima==2.0.4
holidays==0.35

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-grpc==0.42b0
opentelemetry-exporter-otlp==1.21.0
neural_hive_observability==1.1.0
prometheus-client==0.19.0

# Kubernetes / Argo Workflows
kubernetes==28.1.0

# Utilities
httpx==0.25.2
pyyaml==6.0.1
python-dotenv==1.0.0

# Biblioteca centralizada de modelos preditivos
# Para desenvolvimento local: pip install -e ../../libraries/python/neural_hive_ml
-e ../../libraries/python/neural_hive_ml
